requirements.txt
ultralytics
torch
torchvision
torchaudio
pandas
opencv-python
pyyaml
matplotlib
kaggle
onnx
onnxslim
onnxruntime-gpu
